+++ scientificmangement/src/main/java/com/sicnu/mapper/PaperTeamExamineMapper.java
@@ -11,7 +11,7 @@ import java.util.List;
 @Mapper
 @Repository
 public interface PaperTeamExamineMapper {
-    void addPaperTeamExamineUser(Integer pe_id, Integer user_id,Double contribution);
+    void addPaperTeamExamineUser(Integer pe_id, Integer user_id,Integer contribution);
     void delPaperTeamExamineTeam(Integer pe_id);
     List<PaperTeamExamine> selectPaperTeamExamineById(Integer pe_id);
     List<PaperTeamMap> selectPaperTeamExamineUser(Integer pe_id);
