+++ scientificmangement/src/main/java/com/sicnu/mapper/PatentTeamExamineMapper.java
@@ -11,7 +11,7 @@ import java.util.List;
 @Repository
 public interface PatentTeamExamineMapper {
 
-    void addPatentTeamExamineUser(Integer pe_id, Integer user_id, Double contribution);
+    void addPatentTeamExamineUser(Integer pe_id, Integer user_id, Integer contribution);
     void delPatentTeamExamineTeam(Integer pe_id);
     List<PatentTeamExamine> selectPatentTeamExamineById(Integer pe_id);
     List<PatentTeamMap> selectPatentTeamExamineUser(Integer pe_id);
